+++ src/OrlovMikhail.GraphViz.Writing/Attributes/PenWidth/PenWidthAttributeExtensions.cs
@@ -2,7 +2,7 @@
 {
     public static class PenWidthAttributeExtensions
     {
-        public static IAttrSet PenWidth(this IAttrSet attrSet, double value)
+        public static IAttrSet PenWidth(this IAttrSet attrSet, decimal value)
         {
             PenWidthAttribute a = new PenWidthAttribute(value);
             attrSet.Add(a);
