+++ src/OrlovMikhail.GraphViz.Writing/Attributes/RankSep/NodeSepAttributeExtensions.cs
@@ -2,7 +2,7 @@
 {
     public static class NodeSepAttributeExtensions
     {
-        public static IAttrSet NodeSep(this IAttrSet attrSet, double value)
+        public static IAttrSet NodeSep(this IAttrSet attrSet, decimal value)
         {
             NodeSepAttribute a = new NodeSepAttribute(value);
             attrSet.Add(a);
